refactor: web api 관련 코드 타입 분리 및 JSDoc 작성
